requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.8
psycopg2-binary==2.8.3
python-decouple==3.1
python-http-client==3.2.1
pytz==2019.2
requests==2.22.0
sendgrid==6.1.0
sqlparse==0.3.0
urllib3==1.25.6
